dog_cat_classifier/requirements.txt
tensorflow>=2.10.0
pillow>=9.0.0
numpy>=1.22.0
opencv-python>=4.6.0
streamlit>=1.12.0
